+++ testing/mock/hooks.go
@@ -3,7 +3,7 @@ package mock
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	clienttypes "github.com/cosmos/ibc-go/v7/modules/core/02-client/types"
+	clienttypes "github.com/cosmos/ibc-go/v7/modules/core/02-client/types/exp"
 	"github.com/cosmos/ibc-go/v7/modules/core/exported"
 )
